\name{print.DropoutKillerResult}
\alias{print.DropoutKillerResult}
\alias{print.DropoutKillerDetection}
\alias{print.DropoutKillerMembership}
\title{Print DropoutKiller result objects}
\description{Compact print methods for membership, detection, and end-to-end result objects.}
\usage{
\method{print}{DropoutKillerResult}(x, ...)

\method{print}{DropoutKillerDetection}(x, ...)

\method{print}{DropoutKillerMembership}(x, ...)
}
\arguments{
\item{x}{A DropoutKiller result object of the corresponding class.}
\item{...}{Additional arguments, currently unused.}
}
\value{The input object, invisibly.}
